requirements.txt
streamlit==1.37.1
streamlit-webrtc==0.47.7
aiortc==1.9.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
dlib==19.24.4
face-recognition==1.3.0
